# backend/src/syntara/schemas/execution_plane/openapi.yaml
---
openapi: 3.1.0
info:
  title: Execution Plane API
  description: Execution Plane operator endpoints — observe execution targets and work items.
  version: 0.1.0
paths:
  /api/execution_plane/v1/execution_targets:
    get:
      tags:
        - Execution Plane
      summary: List execution targets
      description: Retrieve registered execution targets.
      operationId: list_execution_targets
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionTargetListResponse'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-app-permission:
        resource: execution_target
        action: read
  /api/execution_plane/v1/work_items:
    get:
      tags:
        - Execution Plane
      summary: List work items
      description: Retrieve work items.
      operationId: list_work_items
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemListResponse'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-app-permission:
        resource: work_item
        action: read
components:
  schemas:
    BackendType:
      type: string
      enum:
        - vanilla_k8s
        - openshell
      title: BackendType
      description: Supported execution target backends.
    ExecutionTarget:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        backend_type:
          $ref: '#/components/schemas/BackendType'
        endpoint:
          type: string
          title: Endpoint
        status:
          $ref: '#/components/schemas/TargetStatus'
          default: registering
        enabled:
          type: boolean
          title: Enabled
          default: true
        labels:
          additionalProperties: true
          type: object
          title: Labels
          default: {}
        created_at:
          type: string
          format: date-time
          title: Created At
        last_ran_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Ran At
      type: object
      required:
        - name
        - backend_type
        - endpoint
        - created_at
      title: ExecutionTarget
      description: A registered compute environment where worker pods run.
    ExecutionTargetListResponse:
      properties:
        resources:
          items:
            $ref: '#/components/schemas/ExecutionTarget'
          type: array
          title: Resources
        next:
          anyOf:
            - type: string
            - type: 'null'
          title: Next
        prev:
          anyOf:
            - type: string
            - type: 'null'
          title: Prev
        total:
          anyOf:
            - type: integer
            - type: 'null'
          title: Total
      type: object
      required:
        - resources
      title: ExecutionTargetListResponse
      description: Paginated list response for ExecutionTarget.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    TargetStatus:
      type: string
      enum:
        - registering
        - validating
        - bootstrapping
        - active
        - degraded
        - failed
      title: TargetStatus
      description: Lifecycle states of an execution target.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    WorkItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        work_correlation_id:
          type: string
          format: uuid
          title: Work Correlation Id
        activity_handle:
          type: string
          title: Activity Handle
        status:
          $ref: '#/components/schemas/WorkItemStatus'
          default: pending
        execution_target_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Execution Target Id
        payload:
          additionalProperties: true
          type: object
          title: Payload
          default: {}
        result:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Result
        created_at:
          type: string
          format: date-time
          title: Created At
        claimed_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Claimed At
        completed_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Completed At
        signaled_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Signaled At
      type: object
      required:
        - work_correlation_id
        - activity_handle
        - created_at
      title: WorkItem
      description: A unit of work written by the Temporal Worker and consumed by the Task Executor.
    WorkItemListResponse:
      properties:
        resources:
          items:
            $ref: '#/components/schemas/WorkItem'
          type: array
          title: Resources
        next:
          anyOf:
            - type: string
            - type: 'null'
          title: Next
        prev:
          anyOf:
            - type: string
            - type: 'null'
          title: Prev
        total:
          anyOf:
            - type: integer
            - type: 'null'
          title: Total
      type: object
      required:
        - resources
      title: WorkItemListResponse
      description: Paginated list response for WorkItem.
    WorkItemStatus:
      type: string
      enum:
        - pending
        - claimed
        - dispatched
        - completed
        - failed
        - cancelled
      title: WorkItemStatus
      description: Lifecycle states of a dispatched work item.
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token authentication
      scheme: bearer
      bearerFormat: JWT